server/requirements.txt
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
feedparser==6.0.11
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==24.1
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
setuptools==75.3.0
sgmllib3k==1.0.0
soupsieve==2.6
starkbank-ecdsa==2.2.0
urllib3==2.2.3
Werkzeug==3.1.2
wheel==0.44.0
